refactor[server/utils] Move getLocationFromTweets to utils folder
