+++ server/request-handler.js
@@ -7,7 +7,7 @@ const db = require('../db/config');
 const mongoose = require('mongoose');
 const https = require('https');
 const { Truck, Event } = require('../db/truckSchema');
-const { getLocation } = require('./getLocationFromTweets');
+const { getLocation } = require('../utils/getLocationFromTweets');
 const { getTruckTwitterInfo, createTruckWithGeoInfo, createOrUpdateDB, getYelpInfo, getFiveTweets, updateDBwithYelpInfo, getTenImages } = require('../utils/updateTruckInfo');
 const { createEventRecord, getEventTwitterInfo, createOrUpdateEvent } = require('../utils/updateEventInfo');
 const { geoCoder, newTruckGeoCoder } = require('../utils/utils');
