+++ utils/updateTruckInfo.js
@@ -1,4 +1,3 @@
-// const getLocationFromTweets = require('./getLocationFromTweets');
 const Truck = require('../db/truckSchema');
 const Twitter = require('twitter');
 const Yelp = require('yelp');
